Fix crash (#453)
Fix #451.

When a user creates a person by assigning a name to a face, then that face becomes the person's thumbnail. If that photo is deleted or renamed, then the face record is deleted from Fotema's database, and when Fotema next tries to load the person it will fail because it expects the face thumbnail path to be present.

This PR allows the face thumbnail to be absent.