chore: Update fotema-rust-faces to use ort 2.x (#558)
The ort project has [yanked all the 1.x versions](https://crates.io/crates/ort/versions) of their library from crates.io. To continue to be able to easily update other dependencies, ort must be updated to 2.x.

An annoying side-effect of this is that the ort library can no longer easily be used by multithreads with the Rayon, which dramatically reduces Fotema's face scanning performance. I think this will have to be solved before the next version of Fotema is released.